Extract source-stage/patch/submit MCP tools out of mcp-server.ts (Wave C batch 7)
mcp-server.ts 3976 -> 2604 lines, a 3-file split (the source-staging/
delivery cluster was too big for one or two files under the 500-line
new-file cap):

- mcp-source-stage-tools.ts (createSourceStageTools): get_sources,
  stage_upload_url, stage_source_file. Needs agentTokenSecret + now
  in addition to resolveAuth/injectChannel (mints stage upload tokens).
- mcp-source-patch-tools.ts (createSourcePatchTools): patch_source_file,
  list_staged_sources, clear_staged_sources. Only resolveAuth/injectChannel.
- mcp-source-submit-tools.ts (createSourceSubmitTools): delete_source_file,
  submit_sources. Needs store in addition to resolveAuth/injectChannel
  (reads/writes submission state directly).

All three declare their own minimal local AuthedJob-shaped interface
(wide where needed, e.g. record.slug/roundGeneration/builder), same
precedent as mcp-round-card-tools.ts, rather than importing
mcp-server.ts's private AuthedJob.
