+++ apps/api/src/agent-surface/mcp-server.ts
@@ -1,5 +1,4 @@
 import type { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
-import { z } from 'zod';
 import { AGENT_CHANNEL_ROUTES, type BuilderKind } from '@gamedevpl/contract';
 import {
   toolOk,
@@ -10,7 +9,6 @@ import {
   channelControlFields,
   pendingMessagesFromChannel,
   CREATOR_TEXT_SAFETY,
-  WARNINGS_PROP,
   REPLY_CONTROL,
   PLATFORM_CONNECTOR_ONLY_REASON,
   matchesPlatformConnectorSecret,
@@ -27,10 +25,12 @@ import { createSeedTools } from './mcp-seed-tools.js';
 import { createRoundCardTools } from './mcp-round-card-tools.js';
 import { createGateMediaTools } from './mcp-gate-media-tools.js';
 import { createProposalTools } from './mcp-proposal-tools.js';
+import { createSourceStageTools } from './mcp-source-stage-tools.js';
+import { createSourcePatchTools } from './mcp-source-patch-tools.js';
+import { createSourceSubmitTools } from './mcp-source-submit-tools.js';
 
 export { MCP_VISIBLE_TOOLS };
 import { looksLikeCreatorAgentKey } from './agent-creator-key.js';
-import { canonicalAppBaseUrl } from '../platform/canonical-app-url.js';
 import {
   resolveCreatorAgentKeyForOpenRound,
   resolveCreatorAgentKeyForStart,
@@ -70,19 +70,10 @@ import {
   type AgentTokenAccess,
   type AgentTokenClaims,
 } from './agent-token.js';
-import { DEFAULT_UPLOAD_URL_TTL_SECONDS, mintUploadToken, uploadCurlCommand } from './agent-upload-token.js';
-import { decodeCanonicalBase64Utf8, InvalidBase64Error } from '../platform/canonical-base64.js';
 import { selfBuildDeliveryCap } from '../platform/self-build-delivery-cap.js';
 import type { ManagedUnavailableReason } from './managed-availability.js';
-import {
-  assertDeliverableSourcePath,
-  InvalidUploadError,
-  MAX_UPLOAD_FILES,
-  type GamesStore,
-} from '../delivery/games-store.js';
+import { type GamesStore } from '../delivery/games-store.js';
 import { deriveGateStatusString, readGateVerdict } from '../delivery/gate-verdict.js';
-import { gameManifestHint } from './game-manifest-hint.js';
-import { largeSourceFileHint, moduleSizeWarnings } from '../creation/module-size.js';
 import type { GcsObjectStore } from '../delivery/gcs-sign.js';
 import {
   assertMcpSessionKeyUnexpired,
@@ -166,9 +157,6 @@ const INVALID_START_WINDOW_MS = 60 * 60 * 1000;
 
 /** Hard body ceiling for MCP POSTs (JSON-RPC framing; screenshots use signed PUT). */
 const MAX_MCP_BODY_BYTES = 2 * 1024 * 1024;
-
-const MAX_SUBMIT_FILES = MAX_UPLOAD_FILES;
-const MAX_STAGE_UPLOAD_BATCH = 50;
 
 const TRANSPORT_SESSION_TTL_MS = 24 * 60 * 60 * 1000;
 const MAX_TRANSPORT_SESSIONS = 10_000;
@@ -1935,83 +1923,7 @@ export async function registerMcpServerRoutes(app: FastifyInstance, options: Mcp
     ...createKitFileTools({ resolveAuth, injectChannel }),
     ...createSeedTools({ resolveAuth, injectChannel }),
 
-    get_sources: {
-      annotations: { title: 'Fetch existing game sources', ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          available: { type: 'boolean', description: 'True means this game has files — continue them.' },
-          origin: {
-            type: ['string', 'null'],
-            description: "'seed' = a generated round-0 draft; 'delivery' = a previous round's sources.",
-          },
-          delivery: { type: ['object', 'null'] },
-          files: {
-            type: 'array',
-            items: { type: 'object', properties: { path: { type: 'string' }, content: { type: 'string' } } },
-          },
-          notes: { type: ['string', 'null'], description: 'Hand-off note from the round-0 draft, when there is one.' },
-          references: {
-            type: 'array',
-            items: { type: 'string' },
-            description: 'Published games the round-0 draft was modelled on, when there is one.',
-          },
-          seedStatus: { type: 'string', description: 'pending = a round-0 draft is still generating; call again.' },
-          ...WARNINGS_PROP,
-        },
-        required: ['available', 'files'],
-      },
-      description:
-        "Fetch this game's current sources — the first call of every round, including the first round. " +
-        'A new game already has files: a generated round-0 draft (origin=seed) whose references and notes come ' +
-        'with it. A later round returns what the previous round delivered (origin=delivery). Either way, continue ' +
-        'those files; never scaffold over them. seedStatus=pending means a draft is still generating — browse the ' +
-        'kit briefly and call this again rather than scaffolding. ' +
-        'When warnings.code=module_too_large, split those oversized game/*.ts modules before adding features. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          version: {
-            type: 'string',
-            description: "Optional. Reserved; the channel returns the job's latest delivery or published version.",
-          },
-        },
-        required: [],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const res = await injectChannel(ctx.request, 'GET', AGENT_CHANNEL_ROUTES.SOURCES, auth.channelToken);
-        const body = res.json() as {
-          error?: string;
-          delivery?: unknown;
-          origin?: 'seed' | 'delivery' | null;
-          files?: Array<{ path: string; content: string }>;
-          notes?: string | null;
-          references?: string[];
-          seedStatus?: string;
-        };
-        if (res.statusCode !== 200) {
-          return toolErr(body.error ?? `sources failed (${res.statusCode})`);
-        }
-        const files = body.files ?? [];
-        const sizeWarnings = moduleSizeWarnings(files);
-        // Files decide this, not a delivery: a round-0 draft is sources too.
-        return toolOk({
-          available: files.length > 0,
-          origin: body.origin ?? (body.delivery ? 'delivery' : null),
-          delivery: body.delivery ?? null,
-          files,
-          ...(body.notes ? { notes: body.notes } : {}),
-          ...(body.references?.length ? { references: body.references } : {}),
-          ...(body.seedStatus ? { seedStatus: body.seedStatus } : {}),
-          ...(sizeWarnings.length ? { warnings: sizeWarnings } : {}),
-        });
-      },
-    },
-
+    ...createSourceStageTools({ resolveAuth, injectChannel, agentTokenSecret, now }),
     ...createExampleTools({ resolveAuth, injectChannel }),
 
     report_progress: {
@@ -2183,972 +2095,8 @@ export async function registerMcpServerRoutes(app: FastifyInstance, options: Mcp
       },
     },
 
-    stage_upload_url: {
-      outputSchema: {
-        type: 'object',
-        properties: {
-          url: { type: 'string' },
-          expiresAt: { type: 'string' },
-          expiresInSeconds: { type: 'number' },
-          path: { type: 'string' },
-          upload: { type: 'string' },
-          uploadScript: { type: 'string' },
-          maxBytes: { type: 'number' },
-          uploads: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: {
-                path: { type: 'string' },
-                url: { type: 'string' },
-                upload: { type: 'string' },
-                expiresAt: { type: 'string' },
-                expiresInSeconds: { type: 'number' },
-                maxBytes: { type: 'number' },
-              },
-              required: ['path', 'url', 'upload', 'expiresAt', 'expiresInSeconds', 'maxBytes'],
-            },
-          },
-        },
-      },
-      // Not READS: each call mints a fresh nonce, so it is neither read-only nor idempotent.
-      annotations: { title: 'Get stage upload URL(s)', ...WRITES },
-      description:
-        'Stage new or fully rewritten source file(s) when you have curl/shell egress. ' +
-        'ALWAYS mint upload URLs in batch: pass `paths: ["file1.ts", "file2.ts", ...]` for multiple files ' +
-        `(up to ${MAX_STAGE_UPLOAD_BATCH} paths per call; split larger sets into batches of at most ${MAX_STAGE_UPLOAD_BATCH}; do NOT make individual parallel calls per file). Pass \`path\` only for a lone single file. ` +
-        'Returns short-lived signed PUT URL(s) — run the returned `upload` one-liner(s) ' +
-        '(curl --upload-file <file> "$url") or `uploadScript`. The file bytes never enter the model; the PUT applies the same ' +
-        'validation as stage_source_file (path allowlist, size caps, module_too_large hint) and returns the ' +
-        'staging receipt with stop/pendingMessages. Then submit_sources({ fromStaged: true, … }). ' +
-        'Use stage_source_file / patch_source_file when you have no shell. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          paths: {
-            type: 'array',
-            items: { type: 'string' },
-            description:
-              'Array of game-relative paths (e.g. ["game.ts", "game/render.ts", "GAME.json"]). MANDATORY FOR MULTIPLE FILES: Always pass changed files in batch (up to 50 paths per call; split into batches if staging more). Never emit individual stage_upload_url calls.',
-          },
-          path: {
-            type: 'string',
-            description:
-              'Single game-relative path (e.g. "game.ts"). Use ONLY when staging a single lone file; if staging multiple files, you MUST use paths instead (up to 50 per batch).',
-          },
-          slug: { type: 'string' },
-        },
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        if (!agentTokenSecret) return toolErr('the MCP build endpoint is not configured');
-        const hasPath = typeof args.path === 'string' && args.path.trim().length > 0;
-        const hasPaths = Array.isArray(args.paths) && args.paths.length > 0;
-        if ((!hasPath && !hasPaths) || (hasPath && hasPaths)) {
-          return toolErr(hasPath ? 'pass either path or paths, not both' : 'path or paths is required');
-        }
-        const rawPaths = (hasPaths ? (args.paths as unknown[]) : [args.path]).filter(
-          (p): p is string => typeof p === 'string' && p.trim().length > 0,
-        );
-        if (rawPaths.length === 0) return toolErr('paths must contain at least one valid path');
-        if (rawPaths.length > MAX_STAGE_UPLOAD_BATCH) {
-          return toolErr(
-            `too many paths in one request (max ${MAX_STAGE_UPLOAD_BATCH}; split into batches of up to ${MAX_STAGE_UPLOAD_BATCH})`,
-          );
-        }
-
-        const validPaths: string[] = [];
-        for (const raw of rawPaths) {
-          const trimmed = raw.trim();
-          try {
-            validPaths.push(assertDeliverableSourcePath(trimmed));
-          } catch (error) {
-            if (error instanceof InvalidUploadError) return toolErr(error.message);
-            throw error;
-          }
-        }
-
-        const slug =
-          typeof args.slug === 'string' && args.slug.trim() ? args.slug.trim() : (auth.record.slug ?? undefined);
-        if (!slug && !auth.record.slug) {
-          return toolErr('slug is required before staging — send the slug from get_brief / start');
-        }
-        if (auth.record.slug && slug && auth.record.slug !== slug) {
-          return toolErr(`this build delivers to ${auth.record.slug}, not ${slug}`);
-        }
-        const generation = auth.record.roundGeneration ?? auth.claims.roundGeneration ?? 1;
-        const ttlSeconds = DEFAULT_UPLOAD_URL_TTL_SECONDS;
-        // One clock read: advertised expiresAt must match the signed exp.
-        const issuedAt = now();
-        const expiresAt = new Date(issuedAt + ttlSeconds * 1000).toISOString();
-
-        if (hasPath) {
-          const path = validPaths[0]!;
-          const token = mintUploadToken(agentTokenSecret, {
-            jobId: auth.issueNumber,
-            roundGeneration: generation,
-            kind: 'stage',
-            path,
-            now: issuedAt,
-            ttlSeconds,
-          });
-          const url = `${canonicalAppBaseUrl()}${AGENT_CHANNEL_ROUTES.SOURCES_STAGE_UPLOAD}?token=${encodeURIComponent(token)}`;
-          return toolOk({
-            url,
-            expiresAt,
-            expiresInSeconds: ttlSeconds,
-            path,
-            upload: uploadCurlCommand(url, path, 'text/plain; charset=utf-8'),
-            maxBytes: 1_000_000,
-          });
-        }
-
-        const uploads = validPaths.map((path) => {
-          const token = mintUploadToken(agentTokenSecret, {
-            jobId: auth.issueNumber,
-            roundGeneration: generation,
-            kind: 'stage',
-            path,
-            now: issuedAt,
-            ttlSeconds,
-          });
-          const url = `${canonicalAppBaseUrl()}${AGENT_CHANNEL_ROUTES.SOURCES_STAGE_UPLOAD}?token=${encodeURIComponent(token)}`;
-          return {
-            path,
-            url,
-            upload: uploadCurlCommand(url, path, 'text/plain; charset=utf-8'),
-            expiresAt,
-            expiresInSeconds: ttlSeconds,
-            maxBytes: 1_000_000,
-          };
-        });
-
-        return toolOk({
-          uploads,
-          uploadScript: uploads.map((u) => u.upload).join(' && '),
-          expiresAt,
-          expiresInSeconds: ttlSeconds,
-        });
-      },
-    },
-
-    stage_source_file: {
-      // Overwrites the same path if staged again, so it is not additive.
-      annotations: { title: 'Stage one source file', ...CONSUMES },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          ok: { type: 'boolean' },
-          path: { type: 'string' },
-          bytes: { type: 'number' },
-          hint: { type: 'string' },
-          staged: {
-            type: 'object',
-            properties: {
-              files: {
-                type: 'array',
-                items: {
-                  type: 'object',
-                  properties: { path: { type: 'string' }, bytes: { type: 'number' } },
-                  required: ['path', 'bytes'],
-                },
-              },
-              totalBytes: { type: 'number' },
-              maxBytes: { type: 'number' },
-              maxFiles: { type: 'number' },
-            },
-            required: ['files', 'totalBytes', 'maxBytes', 'maxFiles'],
-          },
-          ...REPLY_CONTROL,
-        },
-        required: ['ok', 'path', 'bytes', 'staged', 'stop', 'pendingMessages'],
-      },
-      description:
-        'Upload ONE game source file into this round’s staging buffer (full rewrite) via inline content. ' +
-        'Use stage_upload_url + curl --upload-file when you have shell egress — re-emitting file contents ' +
-        'as a tool argument burns output tokens. Use this tool for new files when you have no shell; ' +
-        'for edits to an existing path use patch_source_file so you do not re-emit a whole large file. ' +
-        'For a large tree, staging file-by-file avoids one giant submit_sources files[] payload, which some clients truncate. ' +
-        'Call once per path, then submit_sources({ fromStaged: true, mode, kitEngineRef }). Overwrites the same path if staged again. ' +
-        'After preview_failed / red (warnings.code=must_fix_gate), staging alone does not re-run the gate — you must submit_sources again. ' +
-        'Keep modules modest — if hint warns the file is large, split into cohesive game/*.ts modules. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          path: { type: 'string', description: 'Game-relative path (e.g. game.ts, SPEC.md).' },
-          content: { type: 'string', description: 'File contents (utf8 text, or base64 when encoding=base64).' },
-          encoding: { type: 'string', enum: ['utf8', 'base64'] },
-          slug: { type: 'string' },
-        },
-        required: ['path', 'content'],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const path = typeof args.path === 'string' ? args.path.trim() : '';
-        if (!path) return toolErr('path is required');
-        if (typeof args.content !== 'string') return toolErr('content is required');
-        let content = args.content;
-        if (args.encoding === 'base64') {
-          try {
-            content = decodeCanonicalBase64Utf8(args.content);
-          } catch (error) {
-            if (error instanceof InvalidBase64Error) {
-              return toolErr(`file ${path}: invalid base64 content`);
-            }
-            throw error;
-          }
-        }
-        const slug =
-          typeof args.slug === 'string' && args.slug.trim() ? args.slug.trim() : (auth.record.slug ?? undefined);
-        const res = await injectChannel(ctx.request, 'PUT', AGENT_CHANNEL_ROUTES.SOURCES_STAGE, auth.channelToken, {
-          path,
-          content,
-          ...(slug ? { slug } : {}),
-        });
-        const body = res.json() as {
-          error?: string;
-          accepted?: boolean;
-          rejected?: string;
-          path?: string;
-          bytes?: number;
-          hint?: string;
-          manifestHint?: string;
-          typecheckHint?: string;
-          audioHint?: string;
-          staged?: {
-            files: Array<{ path: string; bytes: number }>;
-            totalBytes: number;
-            maxBytes: number;
-            maxFiles: number;
-          };
-          control?: { stop?: boolean; reason?: string };
-          pending?: Array<{ id: string; text: string; createdAt: string }>;
-        };
-        if (res.statusCode !== 200) {
-          return toolErr(body.error ?? `stage failed (${res.statusCode})`, body);
-        }
-        const hint =
-          body.hint ??
-          (typeof body.bytes === 'number' ? largeSourceFileHint(body.path ?? path, body.bytes, content) : null);
-        const manifestHint = body.manifestHint ?? gameManifestHint(body.path ?? path, content);
-        return toolOk({
-          ok: body.accepted !== false,
-          ...(body.rejected ? { rejected: body.rejected } : {}),
-          path: body.path ?? path,
-          bytes: body.bytes ?? 0,
-          ...(hint ? { hint } : {}),
-          staged: body.staged ?? { files: [], totalBytes: 0, maxBytes: 0, maxFiles: 0 },
-          ...channelControlFields(body, [
-            ...(manifestHint ? [{ code: 'game_manifest_invalid' as const, message: manifestHint }] : []),
-            ...(hint ? [{ code: 'module_too_large' as const, message: hint }] : []),
-            ...(body.typecheckHint ? [{ code: 'typecheck_hint' as const, message: body.typecheckHint }] : []),
-            ...(body.audioHint ? [{ code: 'audio_catalog_hint' as const, message: body.audioHint }] : []),
-          ]),
-          pendingMessages: pendingMessagesFromChannel(body),
-        });
-      },
-    },
-
-    patch_source_file: {
-      // Replaces existing staged content, and a patch can remove lines outright.
-      annotations: { title: 'Edit staged source files', ...CONSUMES },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          ok: { type: 'boolean' },
-          path: { type: 'string' },
-          bytes: { type: 'number' },
-          replacements: { type: 'number' },
-          baseFrom: { type: 'string', enum: ['staged', 'delivery', 'seed'] },
-          files: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: {
-                path: { type: 'string' },
-                bytes: { type: 'number' },
-                replacements: { type: 'number' },
-                baseFrom: { type: 'string', enum: ['staged', 'delivery', 'seed'] },
-              },
-              required: ['path', 'bytes', 'replacements', 'baseFrom'],
-            },
-          },
-          incomplete: { type: 'boolean' },
-          failed: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: {
-                path: { type: 'string' },
-                index: { type: 'number' },
-                error: { type: 'string' },
-              },
-              required: ['path', 'index', 'error'],
-            },
-          },
-          hint: { type: 'string' },
-          staged: {
-            type: 'object',
-            properties: {
-              files: {
-                type: 'array',
-                items: {
-                  type: 'object',
-                  properties: { path: { type: 'string' }, bytes: { type: 'number' } },
-                  required: ['path', 'bytes'],
-                },
-              },
-              totalBytes: { type: 'number' },
-              maxBytes: { type: 'number' },
-              maxFiles: { type: 'number' },
-            },
-            required: ['files', 'totalBytes', 'maxBytes', 'maxFiles'],
-          },
-          ...REPLY_CONTROL,
-        },
-        required: ['ok', 'path', 'bytes', 'replacements', 'baseFrom', 'staged', 'stop', 'pendingMessages'],
-      },
-      description:
-        'Edit existing path(s) in the staging buffer without re-uploading whole files. ' +
-        'Edits a file that already exists (from get_sources, a prior stage, or the seed) without re-uploading it whole — ' +
-        'especially useful for large game/render.ts or game/model.ts files. ' +
-        'PRIMARY FORM: pass old + new (exact unique substring replace), or patches: [{ old, new }, ...] for multiple replacements in one file, ' +
-        'or files: [{ path, old, new } | { path, patches: [{ old, new }] }, ...] to edit several files in one call — no @@ line numbers, no diff format. ' +
-        'With patches[] / files[], replacements apply sequentially per file; ensure earlier replacements do not make a later old snippet ambiguous. ' +
-        'Edits that apply are kept even if later ones miss — retry only failed[] (path + index), do not resend the ones that landed. Honour warnings.code=patch_incomplete. ' +
-        'ALTERNATE: pass path + patch as a unified diff for that single file ' +
-        '("--- a/game/render.ts\\n+++ b/game/render.ts\\n@@\\n context\\n-old\\n+new\\n context\\n"; bare @@ ok). ' +
-        'old must match exactly once; widen the snippet if it is ambiguous. Do not mix files[] with top-level path/old/new/patches/patch. ' +
-        'Then submit_sources({ fromStaged: true, mode, kitEngineRef }); fromStaged overlays onto the latest delivery/seed so you only need the patched paths staged. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          path: {
-            type: 'string',
-            description:
-              'Game-relative path (e.g. game/render.ts). Required for single-file edits. For unified diffs, must match the ---/+++ headers. Omit when passing files[].',
-          },
-          old: {
-            type: 'string',
-            description:
-              'Exact text to find (must appear once). Use old+new unless you need a unified diff (patch). Pass together with new.',
-          },
-          new: {
-            type: 'string',
-            description: 'Replacement text for old (may be empty to delete). Pass together with old.',
-          },
-          patches: {
-            type: 'array',
-            description: 'Array of { old, new } replacement pairs to apply sequentially to this file in one call.',
-            items: {
-              type: 'object',
-              properties: {
-                old: { type: 'string', description: 'Exact text to find.' },
-                new: { type: 'string', description: 'Replacement text.' },
-              },
-              required: ['old', 'new'],
-            },
-          },
-          files: {
-            type: 'array',
-            description:
-              'Edit several files in one call. Each entry is { path, old, new } or { path, patches: [{ old, new }, ...] }. Do not pass top-level path/old/new/patches/patch with files[].',
-            items: {
-              type: 'object',
-              properties: {
-                path: { type: 'string', description: 'Game-relative path.' },
-                old: { type: 'string', description: 'Exact text to find (single replace).' },
-                new: { type: 'string', description: 'Replacement text (single replace).' },
-                patches: {
-                  type: 'array',
-                  description: 'Sequential { old, new } pairs for this file.',
-                  items: {
-                    type: 'object',
-                    properties: {
-                      old: { type: 'string' },
-                      new: { type: 'string' },
-                    },
-                    required: ['old', 'new'],
-                  },
-                },
-              },
-              required: ['path'],
-            },
-          },
-          patch: {
-            type: 'string',
-            description:
-              'Unified diff for this one file only (alternative to old+new, patches, or files[]). Bare `@@` hunks are fine when context matches.',
-          },
-          slug: { type: 'string' },
-        },
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const path = typeof args.path === 'string' ? args.path.trim() : '';
-        const hasFiles = Array.isArray(args.files) && args.files.length > 0;
-        const hasPatch = typeof args.patch === 'string' && args.patch.trim().length > 0;
-        const hasOld = typeof args.old === 'string';
-        const hasNew = typeof args.new === 'string';
-        const hasPatches = Array.isArray(args.patches) && args.patches.length > 0;
-
-        if (hasFiles && (path || hasPatch || hasOld || hasNew || hasPatches)) {
-          return toolErr('pass files[] alone, or a single-file path with old+new / patches[] / patch');
-        }
-        if (!hasFiles && !path) return toolErr('path is required unless files[] is passed');
-        const modes = [hasPatch, hasOld || hasNew, hasPatches].filter(Boolean).length;
-        if (!hasFiles && modes > 1) {
-          return toolErr(
-            'pass either old+new (single exact replace), patches[] (multi-replace), files[], or patch (unified diff)',
-          );
-        }
-        if (hasOld !== hasNew) {
-          return toolErr('old and new must be passed together');
-        }
-        if (!hasFiles && modes === 0) {
-          return toolErr('pass old+new, patches[] (multi-replace), files[], or patch (unified diff)');
-        }
-        const slug =
-          typeof args.slug === 'string' && args.slug.trim() ? args.slug.trim() : (auth.record.slug ?? undefined);
-        const res = await injectChannel(
-          ctx.request,
-          'POST',
-          AGENT_CHANNEL_ROUTES.SOURCES_STAGE_PATCH,
-          auth.channelToken,
-          hasFiles
-            ? { files: args.files, ...(slug ? { slug } : {}) }
-            : {
-                path,
-                ...(hasPatches
-                  ? { patches: args.patches }
-                  : hasPatch
-                    ? { patch: args.patch }
-                    : { old: args.old, new: args.new }),
-                ...(slug ? { slug } : {}),
-              },
-        );
-        const body = res.json() as {
-          error?: string;
-          accepted?: boolean;
-          rejected?: string;
-          path?: string;
-          bytes?: number;
-          replacements?: number;
-          baseFrom?: 'staged' | 'delivery' | 'seed';
-          files?: Array<{
-            path: string;
-            bytes: number;
-            replacements: number;
-            baseFrom: 'staged' | 'delivery' | 'seed';
-          }>;
-          incomplete?: boolean;
-          failed?: Array<{ path: string; index: number; error: string }>;
-          hint?: string;
-          manifestHint?: string;
-          typecheckHint?: string;
-          audioHint?: string;
-          staged?: {
-            files: Array<{ path: string; bytes: number }>;
-            totalBytes: number;
-            maxBytes: number;
-            maxFiles: number;
-          };
-          control?: { stop?: boolean; reason?: string };
-          pending?: Array<{ id: string; text: string; createdAt: string }>;
-        };
-        if (res.statusCode !== 200) {
-          return toolErr(body.error ?? `patch failed (${res.statusCode})`, body);
-        }
-        const hint =
-          body.hint ?? (typeof body.bytes === 'number' ? largeSourceFileHint(body.path ?? path, body.bytes) : null);
-        return toolOk({
-          ok: body.accepted !== false,
-          ...(body.rejected ? { rejected: body.rejected } : {}),
-          path: body.path ?? path,
-          bytes: body.bytes ?? 0,
-          replacements: body.replacements ?? 0,
-          baseFrom: body.baseFrom ?? 'staged',
-          ...(body.files ? { files: body.files } : {}),
-          ...(body.incomplete ? { incomplete: true } : {}),
-          ...(body.failed && body.failed.length > 0 ? { failed: body.failed } : {}),
-          ...(hint ? { hint } : {}),
-          staged: body.staged ?? { files: [], totalBytes: 0, maxBytes: 0, maxFiles: 0 },
-          ...channelControlFields(body, [
-            ...(body.manifestHint ? [{ code: 'game_manifest_invalid' as const, message: body.manifestHint }] : []),
-            ...(hint ? [{ code: 'module_too_large' as const, message: hint }] : []),
-            ...(body.typecheckHint ? [{ code: 'typecheck_hint' as const, message: body.typecheckHint }] : []),
-            ...(body.audioHint ? [{ code: 'audio_catalog_hint' as const, message: body.audioHint }] : []),
-            ...(body.failed && body.failed.length > 0
-              ? [
-                  {
-                    code: 'patch_incomplete' as const,
-                    message:
-                      `${body.failed.length} edit${body.failed.length === 1 ? '' : 's'} did not apply — ` +
-                      'retry only failed[] (path + index). Do not resend edits that already landed.',
-                  },
-                ]
-              : []),
-          ]),
-          pendingMessages: pendingMessagesFromChannel(body),
-        });
-      },
-    },
-
-    list_staged_sources: {
-      annotations: { title: 'List staged source files', ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          files: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: { path: { type: 'string' }, bytes: { type: 'number' } },
-              required: ['path', 'bytes'],
-            },
-          },
-          totalBytes: { type: 'number' },
-          maxBytes: { type: 'number' },
-          maxFiles: { type: 'number' },
-          updatedAt: { type: ['string', 'null'] },
-        },
-        required: ['files', 'totalBytes', 'maxBytes', 'maxFiles', 'updatedAt'],
-      },
-      description:
-        'List paths currently in the staging buffer (no contents). Use after stage_source_file / patch_source_file ' +
-        'to confirm changed paths before submit_sources({ fromStaged: true, … }). ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: { sessionKey: SESSION_KEY_PROP },
-        required: [],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const res = await injectChannel(ctx.request, 'GET', AGENT_CHANNEL_ROUTES.SOURCES_STAGE, auth.channelToken);
-        const body = res.json() as {
-          error?: string;
-          files?: Array<{ path: string; bytes: number }>;
-          totalBytes?: number;
-          maxBytes?: number;
-          maxFiles?: number;
-          updatedAt?: string | null;
-        };
-        if (res.statusCode !== 200) {
-          return toolErr(body.error ?? `list staged failed (${res.statusCode})`, body);
-        }
-        return toolOk({
-          files: body.files ?? [],
-          totalBytes: body.totalBytes ?? 0,
-          maxBytes: body.maxBytes ?? 0,
-          maxFiles: body.maxFiles ?? 0,
-          updatedAt: body.updatedAt ?? null,
-        });
-      },
-    },
-
-    clear_staged_sources: {
-      // Deletes staged files — scratch space, nothing creator-visible lost. Idempotent
-      // like ack_inbox: clearing the same paths twice leaves the same result twice.
-      annotations: { title: 'Clear staged source files', ...CONSUMES, idempotentHint: true },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          ok: { type: 'boolean' },
-          cleared: { type: 'number' },
-          ...REPLY_CONTROL,
-        },
-        required: ['ok', 'cleared', 'stop', 'pendingMessages'],
-      },
-      description:
-        'Clear the staging buffer (all paths, or only paths[]). Use before re-staging a clean tree. ' +
-        'Successful submit_sources({ fromStaged: true }) also clears automatically. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          paths: {
-            type: 'array',
-            items: { type: 'string' },
-            description: 'Optional subset to clear; omit to clear everything.',
-          },
-        },
-        required: [],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const paths = Array.isArray(args.paths)
-          ? args.paths.filter((path): path is string => typeof path === 'string' && path.trim().length > 0)
-          : undefined;
-        const res = await injectChannel(
-          ctx.request,
-          'POST',
-          AGENT_CHANNEL_ROUTES.SOURCES_STAGE_CLEAR,
-          auth.channelToken,
-          paths?.length ? { paths } : {},
-        );
-        const body = res.json() as {
-          error?: string;
-          accepted?: boolean;
-          cleared?: number;
-          control?: { stop?: boolean; reason?: string };
-          pending?: Array<{ id: string; text: string; createdAt: string }>;
-        };
-        if (res.statusCode !== 200) {
-          return toolErr(body.error ?? `clear staged failed (${res.statusCode})`, body);
-        }
-        return toolOk({
-          ok: body.accepted !== false,
-          cleared: body.cleared ?? 0,
-          ...channelControlFields(body),
-          pendingMessages: pendingMessagesFromChannel(body),
-        });
-      },
-    },
-
-    delete_source_file: {
-      annotations: { title: 'Delete one staged source file', ...CONSUMES },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          ok: { type: 'boolean' },
-          path: { type: 'string' },
-          staged: {
-            type: 'object',
-            properties: {
-              files: {
-                type: 'array',
-                items: {
-                  type: 'object',
-                  properties: { path: { type: 'string' }, bytes: { type: 'number' } },
-                  required: ['path', 'bytes'],
-                },
-              },
-              totalBytes: { type: 'number' },
-              maxBytes: { type: 'number' },
-              maxFiles: { type: 'number' },
-            },
-            required: ['files', 'totalBytes', 'maxBytes', 'maxFiles'],
-          },
-          ...REPLY_CONTROL,
-        },
-        required: ['ok', 'path', 'staged', 'stop', 'pendingMessages'],
-      },
-      description:
-        'Explicitly remove path from the delivered game — the opposite of stage_source_file. ' +
-        'stage_source_file({ content: "" }) still delivers a live empty file at that path; this instead drops ' +
-        'the path from the next submit_sources({ fromStaged: true }) delivery entirely, same as if it had ' +
-        'never existed. Use to retire an old game/*.ts module no longer imported anywhere, or to clear a ' +
-        'leftover index.html/GAME.json field back to the platform default — index.html cannot be re-staged ' +
-        '(only removed); GAME.json.howToPlay is the only markup source now. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          path: { type: 'string', description: 'Game-relative path to remove (e.g. game/old-module.ts).' },
-        },
-        required: ['path'],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-        const path = typeof args.path === 'string' ? args.path.trim() : '';
-        if (!path) return toolErr('path is required');
-        const res = await injectChannel(
-          ctx.request,
-          'POST',
-          AGENT_CHANNEL_ROUTES.SOURCES_STAGE_DELETE,
-          auth.channelToken,
-          { path },
-        );
-        const body = res.json() as {
-          error?: string;
-          accepted?: boolean;
-          rejected?: string;
-          path?: string;
-          staged?: {
-            files: Array<{ path: string; bytes: number }>;
-            totalBytes: number;
-            maxBytes: number;
-            maxFiles: number;
-          };
-          control?: { stop?: boolean; reason?: string };
-          pending?: Array<{ id: string; text: string; createdAt: string }>;
-        };
-        if (res.statusCode !== 200) {
-          return toolErr(body.error ?? `delete failed (${res.statusCode})`, body);
-        }
-        return toolOk({
-          ok: body.accepted !== false,
-          ...(body.rejected ? { rejected: body.rejected } : {}),
-          path: body.path ?? path,
-          staged: body.staged ?? { files: [], totalBytes: 0, maxBytes: 0, maxFiles: 0 },
-          ...channelControlFields(body),
-          pendingMessages: pendingMessagesFromChannel(body),
-        });
-      },
-    },
-
-    submit_sources: {
-      annotations: { title: 'Deliver sources to the gate', ...CONSUMES },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          ok: { type: 'boolean' },
-          rejected: { type: 'string' },
-          mode: { type: 'string', enum: ['preview', 'publish'] },
-          deliveryId: { type: ['string', 'null'] },
-          delivery: {
-            type: ['object', 'null'],
-            properties: {
-              slug: { type: 'string' },
-              version: { type: 'string' },
-            },
-          },
-          gateStarted: {
-            type: 'boolean',
-            description:
-              'True when Cloud Build accepted the gate create (HTTP 2xx), with or without a parseable build id. ' +
-              'False means the delivery was stored but the gate did not start — do not assume a preview is assembling.',
-          },
-          buildId: {
-            type: 'string',
-            description:
-              'Cloud Build id when the create response included one (may be absent even when gateStarted is true).',
-          },
-          deliveriesRemaining: { type: ['number', 'null'] },
-          ...REPLY_CONTROL,
-        },
-        required: [
-          'ok',
-          'mode',
-          'deliveryId',
-          'delivery',
-          'gateStarted',
-          'deliveriesRemaining',
-          'stop',
-          'pendingMessages',
-        ],
-      },
-      description:
-        `Deliver game sources. Stage changed paths with stage_source_file / patch_source_file, then pass fromStaged=true ` +
-        `(fromStaged overlays onto the latest delivery/seed — do not re-stage unchanged files). ` +
-        `On kit_outdated: get_kit then fromLatestDelivery=true with the same mode and new kitEngineRef — do NOT re-upload the whole tree. ` +
-        `mode=preview (iterate): TRACE/PLAYTEST not required; runs typecheck→smoke→build; Studio gets a draft. ` +
-        `mode=publish (seal): TRACE.json + PLAYTEST.json required; full gate; only publish green ends the round. ` +
-        `Omitting mode defaults to publish, except with fromLatestDelivery (reuses the previous candidate's lane). ` +
-        `files[{path, content, encoding utf8|base64}] optional when fromStaged/fromLatestDelivery (inline paths override); ≤${MAX_SUBMIT_FILES}; kitEngineRef required. ` +
-        'Subject to delivery cap and filename allowlist. Reply includes stop and pendingMessages. ' +
-        'gateStarted is true when Cloud Build accepted the gate create — not merely when the upload was accepted. ' +
-        'A successful delivery unlocks creator handoff (agentEndedAt); still call end when you will not deliver more (warnings.code=call_end). ' +
-        'Pass summary — one sentence of what changed this delivery; Studio shows it on the build list. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          fromStaged: {
-            type: 'boolean',
-            description:
-              'Assemble staging (stage_source_file / patch_source_file), overlaid on the latest delivery and seed. ' +
-              'Use this for large trees and for one-file patches. When true, files[] may be omitted (or used as path overrides). ' +
-              'Not with fromLatestDelivery.',
-          },
-          fromLatestDelivery: {
-            type: 'boolean',
-            description:
-              'Re-deliver the job’s latest candidate from the store (no re-upload). Use after kit_outdated: ' +
-              'get_kit → submit_sources({ fromLatestDelivery:true, mode, kitEngineRef }). Pass the same mode ' +
-              'as the refused delivery (preview stays preview); if mode is omitted the previous lane is inferred. ' +
-              'Optional files[] overlay only the paths you changed. Not with fromStaged.',
-          },
-          files: {
-            type: 'array',
-            maxItems: MAX_SUBMIT_FILES,
-            items: {
-              type: 'object',
-              properties: {
-                path: { type: 'string' },
-                content: { type: 'string' },
-                encoding: { type: 'string', enum: ['utf8', 'base64'] },
-              },
-              required: ['path', 'content'],
-            },
-          },
-          kitEngineRef: {
-            type: 'string',
-            description: 'Creator Kit engineRef the sources were built against (from get_kit / kit.json).',
-          },
-          mode: {
-            type: 'string',
-            enum: ['preview', 'publish'],
-            description:
-              'preview = iterate without TRACE (Studio draft). publish = sealed candidate (TRACE required). ' +
-              'Default publish when omitted, except fromLatestDelivery reuses the previous candidate lane.',
-          },
-          slug: { type: 'string' },
-          note: { type: 'string' },
-          summary: {
-            type: 'string',
-            description:
-              'One sentence of what changed in this delivery (≤1024 chars). Studio shows it on the build list. ' +
-              'Write what the creator should read — not a file dump. end({ summary }) can refine the same sentence later.',
-          },
-        },
-        required: ['kitEngineRef'],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args);
-        if (!('channelToken' in auth)) return auth;
-
-        const fromStaged = args.fromStaged === true;
-        const fromLatestDelivery = args.fromLatestDelivery === true;
-        if (fromStaged && fromLatestDelivery) {
-          return toolErr('fromStaged and fromLatestDelivery cannot both be true — pick one');
-        }
-        const filesParse = z
-          .array(
-            z.object({
-              path: z.string().trim().min(1).max(120),
-              content: z.string(),
-              encoding: z.enum(['utf8', 'base64']).optional(),
-            }),
-          )
-          .max(MAX_SUBMIT_FILES)
-          .optional()
-          .safeParse(args.files);
-        if (!filesParse.success) {
-          return toolErr(filesParse.error.issues[0]?.message ?? 'invalid files');
-        }
-        const inlineFiles = filesParse.data ?? [];
-        if (!fromStaged && !fromLatestDelivery && inlineFiles.length === 0) {
-          return toolErr(
-            'submit_sources needs files[], fromStaged=true after stage_source_file, or fromLatestDelivery=true. ' +
-              'On kit_outdated: get_kit then submit_sources({ fromLatestDelivery: true, mode, kitEngineRef }). ' +
-              'For large first trees: stage_source_file each path, then fromStaged=true.',
-          );
-        }
-
-        const kitEngineRef = typeof args.kitEngineRef === 'string' ? args.kitEngineRef.trim() : '';
-        if (!kitEngineRef) {
-          return toolErr('kitEngineRef is required — send the engineRef from get_kit / kit.json');
-        }
-
-        // Pass mode through only when the agent set it. Omitting lets the channel infer
-        // the previous candidate's lane for fromLatestDelivery (preview kit_outdated
-        // recovery must not suddenly demand TRACE).
-        const mode = args.mode === 'preview' || args.mode === 'publish' ? args.mode : undefined;
-
-        const decodedFiles: Array<{ path: string; content: string }> = [];
-        for (const file of inlineFiles) {
-          if (file.encoding === 'base64') {
-            try {
-              decodedFiles.push({ path: file.path, content: decodeCanonicalBase64Utf8(file.content) });
-            } catch (error) {
-              if (error instanceof InvalidBase64Error) {
-                return toolErr(`file ${file.path}: invalid base64 content`);
-              }
-              throw error;
-            }
-          } else {
-            decodedFiles.push({ path: file.path, content: file.content });
-          }
-        }
-
-        const slug =
-          typeof args.slug === 'string' && args.slug.trim() ? args.slug.trim() : (auth.record.slug ?? 'game');
-        const summary =
-          typeof args.summary === 'string' && args.summary.trim()
-            ? args.summary.trim()
-            : typeof args.note === 'string' && args.note.trim()
-              ? args.note.trim()
-              : undefined;
-
-        const res = await injectChannel(ctx.request, 'POST', AGENT_CHANNEL_ROUTES.SOURCES, auth.channelToken, {
-          slug,
-          ...(decodedFiles.length ? { files: decodedFiles } : {}),
-          ...(fromStaged ? { fromStaged: true } : {}),
-          ...(fromLatestDelivery ? { fromLatestDelivery: true } : {}),
-          kitEngineRef,
-          ...(mode ? { mode } : {}),
-          ...(summary ? { summary } : {}),
-        });
-        const body = res.json() as {
-          error?: string;
-          reason?: string;
-          accepted?: boolean;
-          rejected?: string;
-          mode?: string;
-          delivery?: { slug: string; version: string };
-          deliveryCap?: number;
-          deliveriesUsed?: number;
-          gateStarted?: boolean;
-          buildId?: string;
-          control?: { stop?: boolean; reason?: string };
-          pending?: Array<{ id: string; text: string; createdAt: string }>;
-        };
-        if (res.statusCode !== 200) {
-          return toolErr(body.error ?? `submit failed (${res.statusCode})`, body);
-        }
-
-        const cap = auth.record.builder === 'self' ? selfBuildDeliveryCap() : null;
-        const used =
-          (await store!.getSubmission(auth.issueNumber))?.roundDeliveryCount ?? auth.record.roundDeliveryCount ?? 0;
-
-        const accepted = body.accepted !== false;
-        const gateStarted = body.gateStarted === true;
-        // ChatGPT-class agents usually submit and stop without calling end. Soft
-        // call_end was ignored; mark ended here so Studio unlocks self→platform
-        // handoff immediately. Further channel writes clear agentEndedAt again.
-        if (accepted && store) {
-          await store.markAgentEnded(auth.issueNumber, undefined, 'submit').catch(() => {});
-        }
-        const warnings: Array<{ code: string; message: string }> = [];
-        if (accepted) {
-          warnings.push({
-            code: 'call_end',
-            message:
-              'Call end when you will not deliver more this round (sets stop:true). ' +
-              'Creator handoff is already unlocked from this submit; without end your session may look finished while still connected. ' +
-              'Call end instead of sitting in a get_gate_verdict loop — Studio shows the gate. ' +
-              'If you need an already-available verdict to keep iterating, call get_gate_verdict once; a pending delivery returns stop:true and ends this run.',
-          });
-          if (!gateStarted) {
-            warnings.push({
-              code: 'gate_not_started',
-              message:
-                'Delivery accepted but the gate did not start (no Cloud Build id). ' +
-                'Do not assume a Studio preview is assembling — retry submit_sources or tell the creator.',
-            });
-          }
-        }
-        return toolOk({
-          ok: accepted,
-          mode: body.mode === 'preview' ? 'preview' : 'publish',
-          ...(body.rejected ? { rejected: body.rejected } : {}),
-          deliveryId: body.delivery?.version ?? null,
-          delivery: body.delivery ?? null,
-          gateStarted,
-          ...(typeof body.buildId === 'string' && body.buildId ? { buildId: body.buildId } : {}),
-          deliveriesRemaining: cap === null ? null : Math.max(0, cap - used),
-          ...channelControlFields(body, warnings),
-          pendingMessages: pendingMessagesFromChannel(body),
-        });
-      },
-    },
+    ...createSourcePatchTools({ resolveAuth, injectChannel }),
+    ...createSourceSubmitTools({ resolveAuth, injectChannel, store }),
 
     end: {
       annotations: { title: 'End (commit) this round', ...CONSUMES, idempotentHint: true },
