+++ eslint-rules/module-boundary-map.mjs
@@ -184,6 +184,9 @@ const FILE_BUCKET = {
   'mcp-round-card-tools': 'agent-surface',
   'mcp-gate-media-tools': 'agent-surface',
   'mcp-proposal-tools': 'agent-surface',
+  'mcp-source-stage-tools': 'agent-surface',
+  'mcp-source-patch-tools': 'agent-surface',
+  'mcp-source-submit-tools': 'agent-surface',
   'mcp-server-discovery': 'agent-surface',
   'mcp-ui': 'agent-surface',
   'mcp-oauth-metadata': 'agent-surface',
